Texture Sheet Refactor
Various changes to the texture sheet ROP:
Changed hard coded frame range internally to point to top level parm
Added render buttons for each stage of the process
Added a script function to check for the material of the incoming node to render.
In the comp the incoming image sequence is scaled down before converting to a mosaic - this fixed an issue where higher res images converted to a mosaic would cause a failure because the mosaic has a hard coded limit of 11000x11000 pixels.
Add a camera resolution override toggle - this will match the camera resolution with the resolution of each frame in the final mosaic.
Converted python to hscript for all parm references.
Most camera parms are now linked to the external camera chosen.