Added a toggle for mobile export to VAT
The user can now choose to export vertex animation textures for mobile. This will add a gamma curve and change the image to 16bit integer.  It's recommended to export the textures as tga.

Added a more robust point check for fluid exports. This should fix issues where the user observes black triangles in the game engine.

Changed the render node and camera input on the texture sheets rop to explicitly take an object. This fixes an issue where the render would produce black frames for the beauty image plane.